Following recent discussions with academic and industry partners, we are adjusting the TIM to focus on CO2 emissions. While we strongly believe in including non-CO2 effects in the model long-term, the details of how and when to include these factors requires more input from our stakeholders as part of a governance model that’s in development. With this change, we are provisionally removing contrails effects from our CO2e estimates but will keep the labeling as “CO2e” in the model to ensure future compatibility.
We believe CO2e factors are critical to include in the model,  given the emphasis on them in the IPCC’s AR6 report. We want to make sure that when we do incorporate them into the model, we have a strong plan to account for time of day and regional variations in contrails’ warming impact. We are committed to providing consumers the most accurate information as they make informed choices about their travel options.

We continue to invest into research and collaborate with leading scientists, NGOs, and partners to better incorporate contrails and other non-GHG impact into our model, and we look forward to sharing updates at a later date.